refactor: code structure and clean codebase
