+++ main.py
@@ -1,16 +1,15 @@
 #!/usr/bin/python3
-# -*- coding: utf-8 -*-
-"""
-Entry point for the wallpaper changer application.
+"""Entry point for WallpaperChanger application.
 
 This script serves as the main entry point for running the wallpaper changer,
 which sets wallpapers based on the day of the week and monitor setup.
 
 Example:
     $ python main.py
+
 """
 
-from src import wallpaper
+from src.cli import run
 
 if __name__ == "__main__":
-    wallpaper.main()
+    run()
