+++ pyproject.toml
@@ -17,7 +17,9 @@ classifiers = [
 ]
 
 [project.scripts]
-wallpaperchanger = "wallpaper:main"
+wallpaperchanger = "src.cli:run"
+wallpaperchanger-init = "src.cli:init_config"
+wallpaperchanger-validate = "src.cli:validate_config"
 
 [project.optional-dependencies]
 dev = [
@@ -35,12 +37,7 @@ requires = ["setuptools>=79", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-package-dir = {"" = "src"}
-py-modules = ["wallpaper"]
-
-[tool.setuptools.packages.find]
-where = ["src"]
-include = ["wallpaper*", "wallpaper_changer*"]
+packages = ["src"]
 
 [tool.coverage.paths]
 source = ["wallpaper"]
