+++ src/wallpaper.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
-"""Automatically sets wallpapers based on configuration.
+"""Core wallpaper selection and setting logic.
 
-This script selects wallpapers from designated directories depending on:
+This module handles wallpaper selection from designated directories based on:
 - Day of the week (workday vs configured holidays)
 - Time of day (light/dark themes)
 - Number of connected monitors
@@ -10,68 +10,23 @@ This script selects wallpapers from designated directories depending on:
 Wallpapers are set using appropriate tools (feh for X11, swaybg for Sway
 on Wayland).
 
-Features:
-- Configurable holidays (not just Sunday)
-- Time-based wallpaper selection (day/night)
-- Work vs Holiday wallpaper distinction
-- Multi-monitor support with alternating wallpaper directories
-- Logging with rotation for troubleshooting
-- Display server detection (X11/Wayland)
-
 Example:
-    >>> from wallpaper_changer import wallpaper
-    >>> wallpaper.main()
+    >>> from src import wallpaper, config
+    >>> cfg = config.load()
+    >>> server = wallpaper.detect_display_server()
+    >>> monitors = wallpaper.get_x11_monitors()
+    >>> paths = wallpaper.select_wallpapers(cfg, 2, False, True)
+    >>> wallpaper.set_x11_wallpaper(paths)
 
 """
 
-import datetime
 import json
 import logging
 import os
 import subprocess
-import sys
-from logging.handlers import RotatingFileHandler
 from pathlib import Path
 
-from src.config_loader import Config, create_default_config, load_config
-from src.state_manager import (
-    StateDict,
-    cleanup_old_entries,
-    get_next_wallpaper,
-    initialize_state,
-    load_state,
-    save_state,
-    update_state,
-)
-
-
-def configure_logging(config: Config) -> None:
-    """Configure rotating logs with timestamps and severity levels.
-
-    Args:
-        config: Application configuration with logging settings
-
-    """
-    log_dir = config.logging.log_dir
-    log_file = log_dir / "main.log"
-
-    # Ensure log directory exists
-    log_dir.mkdir(parents=True, exist_ok=True)
-
-    logger = logging.getLogger()
-    logger.setLevel(config.logging.log_level)
-
-    # Clear any existing handlers
-    logger.handlers.clear()
-
-    handler = RotatingFileHandler(
-        str(log_file),
-        maxBytes=config.logging.max_size_mb * 1024 * 1024,
-        backupCount=config.logging.backup_count,
-    )
-    formatter = logging.Formatter("%(asctime)s - %(levelname)s - %(message)s")
-    handler.setFormatter(formatter)
-    logger.addHandler(handler)
+from src import config, state
 
 
 def detect_display_server() -> str:
@@ -176,20 +131,20 @@ def set_sway_wallpaper(image_paths: list[Path], monitors: list[str]) -> None:
 
 
 def select_wallpapers(
-    config: Config,
+    cfg: config.Config,
     monitor_count: int,
     is_holiday: bool,
     is_daytime: bool,
-    state: StateDict | None = None,
+    state_data: state.StateDict | None = None,
 ) -> list[Path]:
     """Select appropriate wallpapers based on configuration and context.
 
     Args:
-        config: Application configuration
+        cfg: Application configuration
         monitor_count: Number of monitors detected
         is_holiday: Whether today is a configured holiday
         is_daytime: Whether current time is during day hours
-        state: Optional state dictionary for round-robin selection
+        state_data: Optional state dictionary for round-robin selection
 
     Returns:
         List of wallpaper paths for each monitor
@@ -197,7 +152,7 @@ def select_wallpapers(
     """
     # Get appropriate directories based on context
     try:
-        wallpaper_dirs = config.get_wallpaper_dirs(is_holiday, is_daytime)
+        wallpaper_dirs = cfg.get_wallpaper_dirs(is_holiday, is_daytime)
     except ValueError as e:
         logging.error("Failed to get wallpaper directories: %s", e)
         return []
@@ -214,11 +169,11 @@ def select_wallpapers(
         else:
             directory = wallpaper_dirs["primary"]
 
-        # Use get_next_wallpaper for both random and round-robin selection
-        path = get_next_wallpaper(
+        # Use state.next_wallpaper for both random and round-robin selection
+        path = state.next_wallpaper(
             directory,
-            config.image_extensions,
-            state,
+            cfg.image_extensions,
+            state_data,
             used_images,
         )
 
@@ -228,152 +183,3 @@ def select_wallpapers(
             logging.error("Failed to select wallpaper from %s", directory)
 
     return wallpaper_paths
-
-
-def main() -> None:
-    """Run main execution flow to set wallpapers based on configuration."""
-    config_path = Path.home() / ".config" / "wallpaperchanger" / "config.ini"
-
-    # Auto-create default config on first run
-    if not config_path.exists():
-        print(
-            "⚠️  Configuration file not found. "
-            "Creating default configuration...",
-            file=sys.stderr,
-        )
-        try:
-            create_default_config(config_path)
-            print(
-                f"✅ Created default configuration at: {config_path}",
-                file=sys.stderr,
-            )
-            print(
-                "\n⚠️  IMPORTANT: Please edit the configuration file to "
-                "set your wallpaper directories!",
-                file=sys.stderr,
-            )
-            print(f"   Edit: {config_path}", file=sys.stderr)
-            print(
-                "\nThe default configuration is set for a specific user "
-                "setup.",
-                file=sys.stderr,
-            )
-            print(
-                "You MUST update the paths to match your system to avoid "
-                "errors.\n",
-                file=sys.stderr,
-            )
-            print(
-                "After editing the config, run the script again:",
-                file=sys.stderr,
-            )
-            print("  python -m src.wallpaper", file=sys.stderr)
-            sys.exit(0)
-        except Exception as e:
-            print(f"❌ Failed to create default config: {e}", file=sys.stderr)
-            sys.exit(1)
-
-    try:
-        # Load configuration
-        config = load_config(config_path)
-    except ValueError as e:
-        error_msg = f"Configuration error: {e}"
-        print(error_msg, file=sys.stderr)
-        print(
-            f"\nPlease check your configuration file: {config_path}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-    except Exception as e:
-        error_msg = f"Unexpected error loading configuration: {e}"
-        print(error_msg, file=sys.stderr)
-        sys.exit(1)
-
-    # Set up logging
-    configure_logging(config)
-    logging.info("=== Starting wallpaper rotation ===")
-
-    # Load state (if state tracking enabled)
-    state = None
-    if config.state_tracking.enabled:
-        logging.info("State tracking enabled")
-        state = load_state(config.state_tracking.state_file)
-        if state is None:
-            logging.info("Initializing new state file")
-            state = initialize_state()
-
-    # Determine current context
-    now = datetime.datetime.now()
-    weekday = now.weekday()
-    current_time = now.time()
-
-    is_holiday = config.is_holiday(weekday)
-    is_daytime = config.is_daytime(current_time)
-
-    logging.info(
-        "Context: %s, %s, Time: %s",
-        "Holiday" if is_holiday else "Workday",
-        "Day" if is_daytime else "Night",
-        current_time.strftime("%H:%M"),
-    )
-
-    # Detect display server and monitors
-    display_server = detect_display_server()
-
-    if display_server == "wayland":
-        monitors = get_sway_monitors()
-    else:
-        monitors = get_x11_monitors()
-
-    monitor_count = len(monitors)
-    logging.info("Active monitors (%d): %s", monitor_count, monitors)
-
-    if monitor_count == 0:
-        logging.error("No monitors detected!")
-        return
-
-    # Select appropriate wallpapers
-    wallpaper_paths = select_wallpapers(
-        config, monitor_count, is_holiday, is_daytime, state
-    )
-
-    # Validate selections
-    if len(wallpaper_paths) != monitor_count:
-        logging.error(
-            "Selected %d wallpapers for %d monitors!",
-            len(wallpaper_paths),
-            monitor_count,
-        )
-        return
-
-    # Update state before setting wallpapers (if state tracking enabled)
-    if state is not None:
-        update_state(
-            state,
-            wallpaper_paths,
-            monitors,
-        )
-
-        # Auto-cleanup if enabled
-        if config.state_tracking.auto_cleanup:
-            cleanup_old_entries(state, max_age_days=30)
-
-    # Apply wallpapers
-    if display_server == "wayland":
-        set_sway_wallpaper(wallpaper_paths, monitors)
-    elif display_server == "x11":
-        set_x11_wallpaper(wallpaper_paths)
-    else:
-        logging.error("Unsupported display server: %s", display_server)
-        return
-
-    # Save state after successful execution (if state tracking enabled)
-    if state is not None:
-        if save_state(config.state_tracking.state_file, state):
-            logging.info("State saved successfully")
-        else:
-            logging.warning("Failed to save state, continuing anyway")
-
-
-if __name__ == "__main__":
-    main()
